requirements.txt
tensorflow>=2.13.0,<2.16.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
streamlit>=1.28.0